+++ R/test_functions.R
@@ -336,10 +336,9 @@ metafunction <- function(data, k_2 = 0.5, k_3 = 0.2, epsilon = NULL) {
   coefD1 <- sample(coefficients, k)
   d2 <- t(utils::combn(1:k, 2))
   set.seed(epsilon)
-  d2M <- d2[sample(nrow(d2), size = ceiling(k * k_2), replace = FALSE), ]
-  sample.size.d2M <- ifelse(is.vector(d2M) == TRUE, 1, nrow(d2M))
+  d2M <- d2[sample(nrow(d2), size = ceiling(k * k_2), replace = FALSE), , drop = FALSE]
   set.seed(epsilon)
-  coefD2 <- sample(coefficients, sample.size.d2M, replace = TRUE)
+  coefD2 <- sample(coefficients, nrow(d2M), replace = TRUE)
   d3 <- t(utils::combn(1:k, 3))
   set.seed(epsilon)
   size.d3 <- ifelse(nrow(d3) == 1, 1, ceiling(k * k_3))
@@ -350,20 +349,8 @@ metafunction <- function(data, k_2 = 0.5, k_3 = 0.2, epsilon = NULL) {
   coefD3 <- sample(coefficients, sample.size, replace = TRUE)
   output <- sapply(seq_along(all_functions), function(x) function_list[[all_functions[x]]](data[, x]))
   y1 <- Rfast::rowsums(mmult(output, coefD1))
-
-  if (is.vector(d2M) == TRUE) {
-    y2 <- sum(output[, d2M[1]] *  output[, d2M[2]] * coefD2)
-
-  } else {
-    y2 <- Rfast::rowsums(mmult(output[, d2M[, 1]] *  output[, d2M[, 2]], coefD2))
-  }
-
-  if (is.vector(d3M) == TRUE) {
-    y3 <- sum(output[, d3M[1]] *  output[, d3M[2]] * output[, d3M[3]] * coefD3)
-
-  } else {
-    y3 <- Rfast::rowsums(mmult(output[, d3M[, 1]] *  output[, d3M[, 2]] * output[, d3M[, 3]], coefD3))
-  }
+  y2 <- Rfast::rowsums(mmult(output[, d2M[, 1]] *  output[, d2M[, 2]], coefD2))
+  y3 <- Rfast::rowsums(mmult(output[, d3M[, 1]] *  output[, d3M[, 2]] * output[, d3M[, 3]], coefD3))
   Y <- y1 + y2 + y3
   return(Y)
 }